requirements.txt
Flask==3.0.0
pymysql==1.1.0
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
prophet==1.1.5
matplotlib==3.8.2
seaborn==0.13.0
Jinja2==3.1.2
python-dotenv==1.0.0
statsmodels==0.14.0
plotly==5.17.0
dash==2.14.1
werkzeug==3.0.1
click==8.1.7
itsdangerous==2.1.2
MarkupSafe==2.1.3
Pillow==10.1.0
joblib==1.3.2
scipy==1.11.4
pytz==2023.3.post1
holidays==0.37
cmdstanpy==1.2.0